Create Mongolian, Hungarian expressions quiz
A Mongolian-Hungarian language learning quiz where the user has to translate Mongolian words into Hungarian, and it also considers synonyms in the evaluation of correctness. It calculates the percentage of correct answers.
Prints a summary with the total number of questions, the number of correct answers, and the percentage of correct answers.